+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Paths for derivations in xpubs is a list and no longer a string.
 - Get transactions with witness data in segwit networks.
+- Add UTXO cache in memory.
 
 ## 0.4.2
 ### Removed